villes

+++ main.cpp
@@ -34,7 +34,7 @@ int main()
     
     for (const auto &row : content)
     {
-        citynumber.push_back(static_cast<int>(row[0])); // Première colonne : indices
+        citynumber.push_back(row[0]); // Première colonne : indices
         coord.push_back(Point(row[1], row[2])); // Deuxième et troisième colonnes : coordonnées x et y
     }
     //création de l'objet ville
